// lib/types/badges.ts
enum Badges {
  HelloWorld = 1 << 0, // Awarded to everyone because why not!
  Community = 1 << 1, // Awarded for meaningful contributions to the community.
  Placeholder1 = 1 << 2,
  Placeholder2 = 1 << 3,
  Placeholder3 = 1 << 4,
  Placeholder4 = 1 << 5,
  Placeholder5 = 1 << 6,
  Placeholder6 = 1 << 7,
  Placeholder7 = 1 << 8,
  Placeholder8 = 1 << 9,
  CodeContributor = 1 << 10,

  NothingToSee = 1 << 30
}

export default Badges;